+++ map-auth-server/src/main/java/com/nmz/mapauthserver/mapper/SysMenuMapper.java
@@ -13,9 +13,9 @@ import java.util.List;
  * @Date: 2023/9/28-20:22
  */
 @Repository
-public interface SysMenuMapper extends CrudRepository<SysMenuEntity, Integer>, QuerydslPredicateExecutor<SysMenuEntity> {
+public interface SysMenuMapper extends CrudRepository<SysMenuEntity, Long>, QuerydslPredicateExecutor<SysMenuEntity> {
 
 
-    List<SysMenuEntity> findByParentId(Integer parentId);
+    List<SysMenuEntity> findByParentId(Long parentId);
 
 }
